+++ src/application/iced_settings/iced_settings.rs
@@ -1,0 +1,29 @@
+use iced::window::settings::PlatformSpecific;
+use iced::window::Position;
+use iced::{Settings, Size};
+
+pub fn settings() -> Settings {
+    Settings::default()
+}
+
+pub fn windows_settings() -> iced::window::Settings {
+    let icon = iced::window::icon::from_file_data(
+        include_bytes!("../../../images/logo.png"),
+        Some(iced::advanced::graphics::image::image_rs::ImageFormat::Png),
+    )
+    .unwrap();
+    iced::window::Settings {
+        size: Size::new(1080.0, 720.0),
+        position: Position::Centered,
+        min_size: Some(Size::new(960.0, 544.0)),
+        max_size: None,
+        visible: true,
+        resizable: true,
+        decorations: true,
+        transparent: false,
+        level: iced::window::Level::Normal,
+        icon: Some(icon),
+        platform_specific: PlatformSpecific::default(),
+        exit_on_close_request: true,
+    }
+}
