+++ src/application/iced_settings/mod.rs
@@ -1,0 +1,1 @@
+pub mod iced_settings;
