+++ src/application/mod.rs
@@ -1,0 +1,2 @@
+pub mod application;
+pub mod iced_settings;
